+++ src/components/ids-render-loop/ids-render-loop-common.ts
@@ -1,8 +1,0 @@
-/**
- * Gets an accurate, current timestamp from the system.
- * @private
- * @returns {number} a current timestamp
- */
-export function timestamp(): number {
-  return window.performance.now();
-}