+++ src/mixins/ids-ripple-mixin/ids-ripple-mixin.ts
@@ -1,7 +1,6 @@
-import renderLoop from '../../components/ids-render-loop/ids-render-loop-global';
-import IdsRenderLoopItem from '../../components/ids-render-loop/ids-render-loop-item';
 import { attributes } from '../../core/ids-attributes';
 import { stringToBool } from '../../utils/ids-string-utils/ids-string-utils';
+import { waitForAnimationEnd } from '../../utils/ids-dom-utils/ids-dom-utils';
 
 const IdsRippleMixin = (superclass: any) => class extends superclass {
   // HTMLElement containing ripple, typically component container
@@ -9,9 +8,6 @@ const IdsRippleMixin = (superclass: any) => class extends superclass {
 
   // Radius of ripple, defaults to 50
   rippleRadius = 50;
-
-  // Timeout for end of ripple effect
-  rippleTimeout?: IdsRenderLoopItem;
 
   constructor() {
     super();
@@ -49,6 +45,12 @@ const IdsRippleMixin = (superclass: any) => class extends superclass {
     super.connectedCallback?.();
   }
 
+  disconnectedCallback() {
+    super.disconnectedCallback?.();
+    this.removeRipples();
+    this.#removeRippleListeners();
+  }
+
   /**
    * @public
    * @param {HTMLElement} rippleTarget Element containing ripple, default to component container
@@ -59,6 +61,14 @@ const IdsRippleMixin = (superclass: any) => class extends superclass {
     this.rippleTarget = rippleTarget || this.container;
     this.rippleRadius = rippleRadius || this.rippleRadius;
     this.#attachRippleListeners();
+  }
+
+  removeRipples(): void {
+    const rippleTarget = this.rippleTarget;
+    const otherRippleEls = rippleTarget?.querySelectorAll('.ripple-effect');
+    otherRippleEls?.forEach((rippleEl) => {
+      rippleEl.remove();
+    });
   }
 
   /**
@@ -113,15 +123,11 @@ const IdsRippleMixin = (superclass: any) => class extends superclass {
    * @param {number} y the Y coordinate
    * @returns {void}
    */
-  createRipple(x: number | undefined, y: number | undefined) {
+  async createRipple(x: number | undefined, y: number | undefined) {
     if (this.noRipple || this.disabled) return;
 
-    // Remove pre-existing ripples
+    this.removeRipples();
     const rippleTarget = this.rippleTarget;
-    const otherRippleEls = rippleTarget?.querySelectorAll('.ripple-effect');
-    otherRippleEls?.forEach((rippleEl) => {
-      rippleEl.remove();
-    });
 
     // Create ripple element
     const rippleEl = document.createElement('span');
@@ -139,19 +145,25 @@ const IdsRippleMixin = (superclass: any) => class extends superclass {
     rippleEl.style.top = `${btnOffsets.y}px`;
     rippleEl.classList.add('animating');
 
-    // Remove pre-existing ripple timeouts
-    if (this.rippleTimeout) {
-      this.rippleTimeout.destroy(true);
-    }
+    await waitForAnimationEnd(rippleEl, 'ripple-animation');
 
-    // After a short time, remove the ripple effect
-    this.rippleTimeout = renderLoop.register(new IdsRenderLoopItem({
-      duration: 1200,
-      timeoutCallback() {
-        rippleTarget?.classList.remove('is-rippling');
-        rippleEl.remove();
+    rippleTarget?.classList.remove('is-rippling');
+    rippleEl.remove();
+    this.#triggerRippleEndEvent(rippleEl);
+  }
+
+  /**
+   * Dispatches a `rippleend` event on the host element that can notify parent Web Components
+   * that a ripple animation has completed.
+   * @param {HTMLElement} rippleEl reference to the removed ripple element
+   * @returns {void}
+   */
+  #triggerRippleEndEvent(rippleEl: HTMLElement) {
+    this.triggerEvent('rippleend', this, {
+      detail: {
+        element: rippleEl
       }
-    }));
+    });
   }
 
   /**